[S07P31C209-179][BE-정지원] feat : maxId 가져오는 postRepository 로직 추가
Former-commit-id: 0a477cf5b9fce16ec05d5b5f50eff8ecc1b2ad79

+++ testCloud/src/main/java/com/example/testcloud/Service/Impl/PostServiceImpl.java
@@ -52,7 +52,7 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public Boolean deletePost(Integer postId) {
-        Integer res = postRepository.deleteByPostId(postId);
+        Long res = postRepository.deleteByPostId(postId);
         if(res == 0){
             return false;
         } else{
